+++ tests/AdapterTest.php
@@ -6,15 +6,15 @@ class AdapterTest extends AdapterTestCase
 {
     public function testGetAndSet()
     {
-        $return = $this->cache->set('key', 'value');
+        $return = $this->cache->set('test key', 'value');
 
         $this->assertEquals(true, $return);
-        $this->assertEquals('value', $this->cache->get('key'));
+        $this->assertEquals('value', $this->cache->get('test key'));
     }
 
     public function testGetFail()
     {
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
     }
 
     public function testGetNonReferential()
@@ -23,19 +23,19 @@ class AdapterTest extends AdapterTestCase
 
         $object = new \StdClass();
         $object->value = 'test';
-        $this->cache->set('key', $object);
+        $this->cache->set('test key', $object);
 
         // clone the object because we'll be messing with it ;)
         $comparison = clone $object;
 
         // changing the object after it's been cached shouldn't affect cache
         $object->value = 'updated-value';
-        $fromCache = $this->cache->get('key');
+        $fromCache = $this->cache->get('test key');
         $this->assertEquals($comparison, $fromCache);
 
         // changing the value we got from cache shouldn't change what's in cache
         $fromCache->value = 'updated-value-2';
-        $fromCache2 = $this->cache->get('key');
+        $fromCache2 = $this->cache->get('test key');
         $this->assertNotEquals($comparison, $fromCache);
         $this->assertEquals($comparison, $fromCache2);
     }
@@ -43,7 +43,7 @@ class AdapterTest extends AdapterTestCase
     public function testGetMulti()
     {
         $items = array(
-            'key' => 'value',
+            'test key' => 'value',
             'key2' => 'value2',
         );
 
@@ -60,25 +60,25 @@ class AdapterTest extends AdapterTestCase
 
     public function testGetNoCasTokens()
     {
-        $this->cache->get('key', $token);
+        $this->cache->get('test key', $token);
         $this->assertEquals(null, $token);
 
-        $this->cache->getMulti(array('key'), $tokens);
+        $this->cache->getMulti(array('test key'), $tokens);
         $this->assertEquals(array(), $tokens);
     }
 
     public function testGetCasTokensFromFalse()
     {
         // 'false' is also a value, with a token
-        $return = $this->cache->set('key', false);
+        $return = $this->cache->set('test key', false);
 
         $this->assertEquals(true, $return);
 
-        $this->assertEquals(false, $this->cache->get('key', $token));
+        $this->assertEquals(false, $this->cache->get('test key', $token));
         $this->assertNotNull($token);
 
-        $this->assertEquals(array('key' => false), $this->cache->getMulti(array('key'), $tokens));
-        $this->assertNotNull($tokens['key']);
+        $this->assertEquals(array('test key' => false), $this->cache->getMulti(array('test key'), $tokens));
+        $this->assertNotNull($tokens['test key']);
     }
 
     public function testGetCasTokensOverridesTokenValue()
@@ -86,33 +86,33 @@ class AdapterTest extends AdapterTestCase
         $token = 'some-value';
         $tokens = array('some-value');
 
-        $this->assertEquals(false, $this->cache->get('key', $token));
+        $this->assertEquals(false, $this->cache->get('test key', $token));
         $this->assertEquals(null, $token);
 
-        $this->assertEquals(array(), $this->cache->getMulti(array('key'), $tokens));
+        $this->assertEquals(array(), $this->cache->getMulti(array('test key'), $tokens));
         $this->assertEquals(array(), $tokens);
     }
 
     public function testSetExpired()
     {
-        $return = $this->cache->set('key', 'value', time() - 2);
+        $return = $this->cache->set('test key', 'value', time() - 2);
         $this->assertEquals(true, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
 
         // test if we can add to, but not replace or touch an expired value; it
         // should be treated as if the value doesn't exist)
-        $return = $this->cache->replace('key', 'value');
+        $return = $this->cache->replace('test key', 'value');
         $this->assertEquals(false, $return);
-        $return = $this->cache->touch('key', time() + 2);
+        $return = $this->cache->touch('test key', time() + 2);
         $this->assertEquals(false, $return);
-        $return = $this->cache->add('key', 'value');
+        $return = $this->cache->add('test key', 'value');
         $this->assertEquals(true, $return);
     }
 
     public function testSetMulti()
     {
         $items = array(
-            'key' => 'value',
+            'test key' => 'value',
             'key2' => 'value2',
         );
 
@@ -120,14 +120,14 @@ class AdapterTest extends AdapterTestCase
 
         $expect = array_fill_keys(array_keys($items), true);
         $this->assertEquals($expect, $return);
-        $this->assertEquals('value', $this->cache->get('key'));
+        $this->assertEquals('value', $this->cache->get('test key'));
         $this->assertEquals('value2', $this->cache->get('key2'));
     }
 
     public function testSetMultiExpired()
     {
         $items = array(
-            'key' => 'value',
+            'test key' => 'value',
             'key2' => 'value2',
         );
 
@@ -135,18 +135,18 @@ class AdapterTest extends AdapterTestCase
 
         $expect = array_fill_keys(array_keys($items), true);
         $this->assertEquals($expect, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
         $this->assertEquals(false, $this->cache->get('key2'));
     }
 
     public function testDelete()
     {
-        $this->cache->set('key', 'value');
+        $this->cache->set('test key', 'value');
 
-        $return = $this->cache->delete('key');
+        $return = $this->cache->delete('test key');
 
         $this->assertEquals(true, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
 
         // delete non-existing key
         $return = $this->cache->delete('key2');
@@ -158,7 +158,7 @@ class AdapterTest extends AdapterTestCase
     public function testDeleteMulti()
     {
         $items = array(
-            'key' => 'value',
+            'test key' => 'value',
             'key2' => 'value2',
         );
 
@@ -167,7 +167,7 @@ class AdapterTest extends AdapterTestCase
 
         $expect = array_fill_keys(array_keys($items), true);
         $this->assertEquals($expect, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
         $this->assertEquals(false, $this->cache->get('key2'));
 
         // delete all non-existing key (they've been deleted already)
@@ -175,170 +175,170 @@ class AdapterTest extends AdapterTestCase
 
         $expect = array_fill_keys(array_keys($items), false);
         $this->assertEquals($expect, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
         $this->assertEquals(false, $this->cache->get('key2'));
 
         // delete existing & non-existing key
-        $this->cache->set('key', 'value');
+        $this->cache->set('test key', 'value');
         $return = $this->cache->deleteMulti(array_keys($items));
 
         $expect = array_fill_keys(array_keys($items), false);
-        $expect['key'] = true;
+        $expect['test key'] = true;
         $this->assertEquals($expect, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
         $this->assertEquals(false, $this->cache->get('key2'));
     }
 
     public function testAdd()
     {
-        $return = $this->cache->add('key', 'value');
+        $return = $this->cache->add('test key', 'value');
 
         $this->assertEquals(true, $return);
-        $this->assertEquals('value', $this->cache->get('key'));
+        $this->assertEquals('value', $this->cache->get('test key'));
     }
 
     public function testAddFail()
     {
-        $this->cache->set('key', 'value');
-        $return = $this->cache->add('key', 'value-2');
+        $this->cache->set('test key', 'value');
+        $return = $this->cache->add('test key', 'value-2');
 
         $this->assertEquals(false, $return);
-        $this->assertEquals('value', $this->cache->get('key'));
+        $this->assertEquals('value', $this->cache->get('test key'));
     }
 
     public function testAddExpired()
     {
-        $return = $this->cache->add('key', 'value', time() - 2);
+        $return = $this->cache->add('test key', 'value', time() - 2);
 
         $this->assertEquals(true, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
     }
 
     public function testReplace()
     {
-        $this->cache->set('key', 'value');
-        $return = $this->cache->replace('key', 'value-2');
+        $this->cache->set('test key', 'value');
+        $return = $this->cache->replace('test key', 'value-2');
 
         $this->assertEquals(true, $return);
-        $this->assertEquals('value-2', $this->cache->get('key'));
+        $this->assertEquals('value-2', $this->cache->get('test key'));
     }
 
     public function testReplaceFail()
     {
-        $return = $this->cache->replace('key', 'value');
+        $return = $this->cache->replace('test key', 'value');
 
         $this->assertEquals(false, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
     }
 
     public function testReplaceExpired()
     {
-        $this->cache->set('key', 'value');
-        $return = $this->cache->replace('key', 'value', time() - 2);
+        $this->cache->set('test key', 'value');
+        $return = $this->cache->replace('test key', 'value', time() - 2);
 
         $this->assertEquals(true, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
     }
 
     public function testReplaceSameValue()
     {
-        $this->cache->set('key', 'value');
-        $return = $this->cache->replace('key', 'value');
+        $this->cache->set('test key', 'value');
+        $return = $this->cache->replace('test key', 'value');
 
         $this->assertEquals(true, $return);
     }
 
     public function testCas()
     {
-        $this->cache->set('key', 'value');
+        $this->cache->set('test key', 'value');
 
         // token via get()
-        $this->cache->get('key', $token);
-        $return = $this->cache->cas($token, 'key', 'updated-value');
+        $this->cache->get('test key', $token);
+        $return = $this->cache->cas($token, 'test key', 'updated-value');
 
         $this->assertEquals(true, $return);
-        $this->assertEquals('updated-value', $this->cache->get('key'));
+        $this->assertEquals('updated-value', $this->cache->get('test key'));
 
         // token via getMulti()
-        $this->cache->getMulti(array('key'), $tokens);
-        $token = $tokens['key'];
-        $return = $this->cache->cas($token, 'key', 'updated-value-2');
+        $this->cache->getMulti(array('test key'), $tokens);
+        $token = $tokens['test key'];
+        $return = $this->cache->cas($token, 'test key', 'updated-value-2');
 
         $this->assertEquals(true, $return);
-        $this->assertEquals('updated-value-2', $this->cache->get('key'));
+        $this->assertEquals('updated-value-2', $this->cache->get('test key'));
     }
 
     public function testCasFail()
     {
-        $this->cache->set('key', 'value');
+        $this->cache->set('test key', 'value');
 
         // get CAS token
-        $this->cache->get('key', $token);
+        $this->cache->get('test key', $token);
 
         // write something else to the same key in the meantime
-        $this->cache->set('key', 'updated-value');
+        $this->cache->set('test key', 'updated-value');
 
         // attempt CAS, which should now fail (token no longer valid)
-        $return = $this->cache->cas($token, 'key', 'updated-value-2');
+        $return = $this->cache->cas($token, 'test key', 'updated-value-2');
 
         $this->assertEquals(false, $return);
-        $this->assertEquals('updated-value', $this->cache->get('key'));
+        $this->assertEquals('updated-value', $this->cache->get('test key'));
     }
 
     public function testCasFail2()
     {
-        $this->cache->set('key', 'value');
+        $this->cache->set('test key', 'value');
 
         // get CAS token
-        $this->cache->get('key', $token);
+        $this->cache->get('test key', $token);
 
         // delete that key in the meantime
-        $this->cache->delete('key');
+        $this->cache->delete('test key');
 
         // attempt CAS, which should now fail (token no longer valid)
-        $return = $this->cache->cas($token, 'key', 'updated-value-2');
+        $return = $this->cache->cas($token, 'test key', 'updated-value-2');
 
         $this->assertEquals(false, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
     }
 
     public function testCasExpired()
     {
-        $this->cache->set('key', 'value');
+        $this->cache->set('test key', 'value');
 
         // token via get()
-        $this->cache->get('key', $token);
-        $return = $this->cache->cas($token, 'key', 'updated-value', time() - 2);
+        $this->cache->get('test key', $token);
+        $return = $this->cache->cas($token, 'test key', 'updated-value', time() - 2);
 
         $this->assertEquals(true, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
     }
 
     public function testCasSameValue()
     {
-        $this->cache->set('key', 'value');
-        $this->cache->get('key', $token);
-        $return = $this->cache->cas($token, 'key', 'value');
+        $this->cache->set('test key', 'value');
+        $this->cache->get('test key', $token);
+        $return = $this->cache->cas($token, 'test key', 'value');
 
         $this->assertEquals(true, $return);
     }
 
     public function testCasNoOriginalValue()
     {
-        $this->cache->get('key', $token);
-        $return = $this->cache->cas($token, 'key', 'value');
+        $this->cache->get('test key', $token);
+        $return = $this->cache->cas($token, 'test key', 'value');
 
         $this->assertEquals(false, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
     }
 
     public function testIncrement()
     {
         // set initial value
-        $return = $this->cache->increment('key', 1, 1);
+        $return = $this->cache->increment('test key', 1, 1);
 
         $this->assertEquals(1, $return);
-        $this->assertEquals(1, $this->cache->get('key'));
+        $this->assertEquals(1, $this->cache->get('test key'));
 
         $return = $this->cache->increment('key2', 1, 0);
 
@@ -346,52 +346,52 @@ class AdapterTest extends AdapterTestCase
         $this->assertEquals(0, $this->cache->get('key2'));
 
         // increment
-        $return = $this->cache->increment('key', 1, 1);
+        $return = $this->cache->increment('test key', 1, 1);
 
         $this->assertEquals(2, $return);
-        $this->assertEquals(2, $this->cache->get('key'));
+        $this->assertEquals(2, $this->cache->get('test key'));
     }
 
     public function testIncrementFail()
     {
-        $return = $this->cache->increment('key', -1, 0);
+        $return = $this->cache->increment('test key', -1, 0);
         $this->assertEquals(false, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
 
-        $return = $this->cache->increment('key', 5, -2);
+        $return = $this->cache->increment('test key', 5, -2);
         $this->assertEquals(false, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
 
         // non-numeric value in cache
-        $this->cache->set('key', 'value');
-        $return = $this->cache->increment('key', 1, 1);
+        $this->cache->set('test key', 'value');
+        $return = $this->cache->increment('test key', 1, 1);
         $this->assertEquals(false, $return);
-        $this->assertEquals('value', $this->cache->get('key'));
+        $this->assertEquals('value', $this->cache->get('test key'));
     }
 
     public function testIncrementExpired()
     {
         // set initial value
-        $return = $this->cache->increment('key', 1, 1, time() - 2);
+        $return = $this->cache->increment('test key', 1, 1, time() - 2);
 
         $this->assertEquals(1, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
 
         // set initial value (not expired) & increment (expired)
-        $this->cache->increment('key', 1, 1);
-        $return = $this->cache->increment('key', 1, 1, time() - 2);
+        $this->cache->increment('test key', 1, 1);
+        $return = $this->cache->increment('test key', 1, 1, time() - 2);
 
         $this->assertEquals(2, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
     }
 
     public function testDecrement()
     {
         // set initial value
-        $return = $this->cache->decrement('key', 1, 1);
+        $return = $this->cache->decrement('test key', 1, 1);
 
         $this->assertEquals(1, $return);
-        $this->assertEquals(1, $this->cache->get('key'));
+        $this->assertEquals(1, $this->cache->get('test key'));
 
         $return = $this->cache->decrement('key2', 1, 0);
 
@@ -399,75 +399,75 @@ class AdapterTest extends AdapterTestCase
         $this->assertEquals(0, $this->cache->get('key2'));
 
         // decrement
-        $return = $this->cache->decrement('key', 1, 1);
+        $return = $this->cache->decrement('test key', 1, 1);
 
         $this->assertEquals(0, $return);
-        $this->assertEquals(0, $this->cache->get('key'));
+        $this->assertEquals(0, $this->cache->get('test key'));
 
         // decrement again (can't go below 0)
-        $return = $this->cache->decrement('key', 1, 1);
+        $return = $this->cache->decrement('test key', 1, 1);
 
         $this->assertEquals(0, $return);
-        $this->assertEquals(0, $this->cache->get('key'));
+        $this->assertEquals(0, $this->cache->get('test key'));
     }
 
     public function testDecrementFail()
     {
-        $return = $this->cache->decrement('key', -1, 0);
+        $return = $this->cache->decrement('test key', -1, 0);
         $this->assertEquals(false, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
 
-        $return = $this->cache->decrement('key', 5, -2);
+        $return = $this->cache->decrement('test key', 5, -2);
         $this->assertEquals(false, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
 
         // non-numeric value in cache
-        $this->cache->set('key', 'value');
-        $return = $this->cache->increment('key', 1, 1);
+        $this->cache->set('test key', 'value');
+        $return = $this->cache->increment('test key', 1, 1);
         $this->assertEquals(false, $return);
-        $this->assertEquals('value', $this->cache->get('key'));
+        $this->assertEquals('value', $this->cache->get('test key'));
     }
 
     public function testDecrementExpired()
     {
         // set initial value
-        $return = $this->cache->decrement('key', 1, 1, time() - 2);
+        $return = $this->cache->decrement('test key', 1, 1, time() - 2);
 
         $this->assertEquals(1, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
 
         // set initial value (not expired) & increment (expired)
-        $this->cache->decrement('key', 1, 1);
-        $return = $this->cache->decrement('key', 1, 1, time() - 2);
+        $this->cache->decrement('test key', 1, 1);
+        $return = $this->cache->decrement('test key', 1, 1, time() - 2);
 
         $this->assertEquals(0, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
     }
 
     public function testTouch()
     {
-        $this->cache->set('key', 'value');
+        $this->cache->set('test key', 'value');
 
         // not yet expired
-        $this->cache->touch('key', time() + 2);
-        $this->assertEquals('value', $this->cache->get('key'));
+        $this->cache->touch('test key', time() + 2);
+        $this->assertEquals('value', $this->cache->get('test key'));
     }
 
     public function testTouchExpired()
     {
-        $this->cache->set('key', 'value');
+        $this->cache->set('test key', 'value');
 
         // expired
-        $this->cache->touch('key', time() - 2);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->cache->touch('test key', time() - 2);
+        $this->assertEquals(false, $this->cache->get('test key'));
     }
 
     public function testFlush()
     {
-        $this->cache->set('key', 'value');
+        $this->cache->set('test key', 'value');
         $return = $this->cache->flush();
 
         $this->assertEquals(true, $return);
-        $this->assertEquals(false, $this->cache->get('key'));
+        $this->assertEquals(false, $this->cache->get('test key'));
     }
 }